refactor: move evaluator and utils to shared/
