+++ src/shared/core/__test__/parser.spec.ts
