+++ src/shared/core/evaluator.ts
