+++ src/shared/core/parser.ts
